Add files via upload
Problem Statement: There are n glasses on the table, all standing upside down. In one move, you are allowed to turn over exactly n-1 of them. Our problem is to determine all the values of n for which the glasses can be turned up, and to outline an algorithm that does this in minimum number of moves.